+++ lib/rounded_loading_button.dart
@@ -5,7 +5,7 @@ import 'package:rxdart/rxdart.dart';
 
 /// States that your button can assume via the controller
 // ignore: public_member_api_docs
-enum LoadingState { idle, loading, success, error }
+enum ButtonState { idle, loading, success, error }
 
 /// Initalize class
 class RoundedLoadingButton extends StatefulWidget {
@@ -40,7 +40,7 @@ class RoundedLoadingButton extends StatefulWidget {
   /// The color of the static icons
   final Color valueColor;
 
-  /// reset the animation after specified duration, 
+  /// reset the animation after specified duration,
   /// use resetDuration parameter to set Duration, defaults to 15 seconds
   final bool resetAfterDuration;
 
@@ -68,6 +68,18 @@ class RoundedLoadingButton extends StatefulWidget {
   /// The color of the button when it is disabled
   final Color? disabledColor;
 
+  /// The icon for the success state
+  final IconData successIcon;
+
+  /// The icon for the failed state
+  final IconData failedIcon;
+
+  /// The success and failed animation curve
+  final Curve completionCurve;
+
+  /// The duration of the success and failed animation
+  final Duration completionDuration;
+
   Duration get _borderDuration {
     return Duration(milliseconds: (duration.inMilliseconds / 2).round());
   }
@@ -93,7 +105,12 @@ class RoundedLoadingButton extends StatefulWidget {
       this.successColor,
       this.resetDuration = const Duration(seconds: 15),
       this.resetAfterDuration = false,
-      this.disabledColor});
+      this.successIcon = Icons.check,
+      this.failedIcon = Icons.close,
+      this.completionCurve = Curves.elasticOut,
+      this.completionDuration = const Duration(milliseconds: 1000),
+      this.disabledColor})
+      : super(key: key);
 
   @override
   State<StatefulWidget> createState() => RoundedLoadingButtonState();
@@ -110,7 +127,7 @@ class RoundedLoadingButtonState extends State<RoundedLoadingButton>
   late Animation _bounceAnimation;
   late Animation _borderAnimation;
 
-  final _state = BehaviorSubject<LoadingState>.seeded(LoadingState.idle);
+  final _state = BehaviorSubject<ButtonState>.seeded(ButtonState.idle);
 
   @override
   Widget build(BuildContext context) {
@@ -127,7 +144,7 @@ class RoundedLoadingButtonState extends State<RoundedLoadingButton>
         height: _bounceAnimation.value,
         child: _bounceAnimation.value > 20
             ? Icon(
-                Icons.check,
+                widget.successIcon,
                 color: widget.valueColor,
               )
             : null);
@@ -143,7 +160,7 @@ class RoundedLoadingButtonState extends State<RoundedLoadingButton>
         height: _bounceAnimation.value,
         child: _bounceAnimation.value > 20
             ? Icon(
-                Icons.close,
+                widget.failedIcon,
                 color: widget.valueColor,
               )
             : null);
@@ -161,13 +178,14 @@ class RoundedLoadingButtonState extends State<RoundedLoadingButton>
         return AnimatedSwitcher(
             duration: Duration(milliseconds: 200),
             child:
-                snapshot.data == LoadingState.loading ? _loader : widget.child);
+                snapshot.data == ButtonState.loading ? _loader : widget.child);
       },
     );
 
     final _btn = ButtonTheme(
         shape: RoundedRectangleBorder(borderRadius: _borderAnimation.value),
         disabledColor: widget.disabledColor,
+        padding: EdgeInsets.all(0),
         child: ElevatedButton(
           style: ElevatedButton.styleFrom(
             onSurface: widget.disabledColor,
@@ -186,9 +204,9 @@ class RoundedLoadingButtonState extends State<RoundedLoadingButton>
     return Container(
         height: widget.height,
         child: Center(
-            child: _state.value == LoadingState.error
+            child: _state.value == ButtonState.error
                 ? _cross
-                : _state.value == LoadingState.success
+                : _state.value == ButtonState.success
                     ? _check
                     : _btn));
   }
@@ -200,15 +218,15 @@ class RoundedLoadingButtonState extends State<RoundedLoadingButton>
     _buttonController =
         AnimationController(duration: widget.duration, vsync: this);
 
-    _checkButtonControler = AnimationController(
-        duration: Duration(milliseconds: 1000), vsync: this);
+    _checkButtonControler =
+        AnimationController(duration: widget.completionDuration, vsync: this);
 
     _borderController =
         AnimationController(duration: widget._borderDuration, vsync: this);
 
     _bounceAnimation = Tween<double>(begin: 0, end: widget.height).animate(
         CurvedAnimation(
-            parent: _checkButtonControler, curve: Curves.elasticOut));
+            parent: _checkButtonControler, curve: widget.completionCurve));
     _bounceAnimation.addListener(() {
       setState(() {});
     });
@@ -238,6 +256,11 @@ class RoundedLoadingButtonState extends State<RoundedLoadingButton>
       setState(() {});
     });
 
+    // There is probably a better way of doing this...
+    _state.stream.listen((event) {
+      widget.controller._state.sink.add(event);
+    });
+
     widget.controller._addListeners(_start, _stop, _success, _error, _reset);
   }
 
@@ -261,31 +284,31 @@ class RoundedLoadingButtonState extends State<RoundedLoadingButton>
   }
 
   _start() {
-    _state.sink.add(LoadingState.loading);
+    _state.sink.add(ButtonState.loading);
     _borderController.forward();
     _buttonController.forward();
-    if(widget.resetAfterDuration) _reset();
+    if (widget.resetAfterDuration) _reset();
   }
 
   _stop() {
-    _state.sink.add(LoadingState.idle);
+    _state.sink.add(ButtonState.idle);
     _buttonController.reverse();
     _borderController.reverse();
   }
 
   _success() {
-    _state.sink.add(LoadingState.success);
+    _state.sink.add(ButtonState.success);
     _checkButtonControler.forward();
   }
 
   _error() {
-    _state.sink.add(LoadingState.error);
+    _state.sink.add(ButtonState.error);
     _checkButtonControler.forward();
   }
 
   _reset() async {
     if (widget.resetAfterDuration) await Future.delayed(widget.resetDuration);
-    _state.sink.add(LoadingState.idle);
+    _state.sink.add(ButtonState.idle);
     _buttonController.reverse();
     _borderController.reverse();
     _checkButtonControler.reset();
@@ -313,6 +336,15 @@ class RoundedLoadingButtonController {
     _errorListener = errorListener;
     _resetListener = resetListener;
   }
+
+  final BehaviorSubject<ButtonState> _state =
+    BehaviorSubject<ButtonState>.seeded(ButtonState.idle);
+    
+  /// A read-only stream of the button state
+  Stream<ButtonState> get stateStream => _state.stream;
+
+  /// Gets the current state
+  ButtonState? get currentState => _state.value;
 
   /// Notify listeners to start the loading animation
   void start() {