+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [2.0.9]
+* Added ability to listen to and query button status from the controller
+
+## [2.0.8]
+* Added success and error icons parameters
+* Added completion curve and duration parameters
+
+## [2.0.7]
+* Updated rxdart constraints 
+* Updated the screenshot
+
+## [2.0.6]
+* Fixed bug with state key
+* Added the auto reset functionality 
+
 ## [2.0.5]
 * Fixed disabled colour issue